Bug fix Hire Smart Project
There's a bug in hire smart project on the website when we click on the button to view complete description of Hire Smart project it is showing description of Artseum Project. I think thats because some details are not added in Hire Smart. I've added them now and most probably the bug is fixed now.